requirements.txt
flair
transformers
pynvml
pyRAPL
pandas
psutil
pymongo
datasets
networkx
preon
python-dotenv
scispacy==0.5.1
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz


--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio